requirements.txt
openpyxl==3.1.5
selenium==4.24.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
cozepy==0.8.0